requirements.txt
#
# Speedr requirements
#
Flask
schedule
speedtest-cli